requirements.txt
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
Flask-WTF==1.1.1
WTForms==3.0.1
twilio==8.8.0
Pillow==10.4.0
bcrypt==4.0.1
psycopg2-binary==2.9.9
cloudinary==1.36.0
folium==0.14.0
branca==0.8.1

# Advanced Features
qrcode==7.4.2
reportlab==4.0.7
python-telegram-bot==20.7
face-recognition==1.3.0
dlib==19.24.6
